Add transformed Kotlin metadata libraries
This commit introduces multiple Kotlin metadata files for various libraries, including Compose and JetBrains components. Additionally, remove obsolete theme-related files from the Compose application, namely `Color.kt`, `Theme.android.kt`, `Theme.desktop.kt`, and `Theme.kt`.

+++ shared/src/desktopMain/kotlin/org/example/shared/data/model/User.kt
@@ -40,7 +40,7 @@ actual data class User(
     val lastLoginAt: String,
     val localId: String,
     val passwordHash: String,
-    val passwordUpdatedAt: Double,
+    val passwordUpdatedAt: Long,
     val photoUrl: String,
     val providerUserInfo: List<ProviderUserInfo>,
     val validSince: String
